requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
beanie==1.29.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
databases==0.9.0
Deprecated==1.2.18
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.11
fastapi-cli==0.0.7
frozenlist==1.6.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
lazy-model==0.2.0
limits==4.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.5
openai==0.28.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
passlib==1.7.4
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.11.2
PyMySQL==1.1.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sqlparse==0.5.3
starlette==0.46.1
timescaledb==0.0.4
toml==0.10.2
tortoise-orm==0.24.2
tqdm==4.67.1
typer==0.15.2
types-pytz==2025.2.0.20250326
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uv==0.6.11
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
